Add disk encryption for SNO
