+++ src/common/components/clusterConfiguration/DiskEncryptionFields/DiskEncryptionValues.ts
@@ -1,0 +1,7 @@
+import { DiskEncryption } from '../../../api';
+
+export interface DiskEncryptionValues {
+  enableDiskEncryptionOnMasters: boolean;
+  enableDiskEncryptionOnWorkers: boolean;
+  diskEncryptionMode: DiskEncryption['mode'];
+}
