+++ src/common/components/clusterWizard/clusterDetailsValidation.ts
@@ -64,10 +64,12 @@ export const getClusterDetailsInitialValues = ({
     enableDiskEncryptionOnWorkers: ['all', 'workers'].includes(
       cluster?.diskEncryption?.enableOn ?? 'none',
     ),
+    enableDiskEncryptionSNO: ['all'].includes(cluster?.diskEncryption?.enableOn ?? 'none'),
     diskEncryptionMode: cluster?.diskEncryption?.mode ?? 'tpmv2',
     diskEncryptionTangServers: cluster?.diskEncryption?.tangServers
       ? parseTangServers(cluster.diskEncryption.tangServers)
       : emptyTangServers(),
+    diskEncryption: cluster?.diskEncryption ?? {},
   };
 };
